% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{epsilon_fn}
\alias{epsilon_fn}
\title{Computer the consistency lower bound}
\usage{
epsilon_fn(df, B)
}
\arguments{
\item{df}{Table of (de)activation rates.}

\item{B}{Number of complementary pairs to draw for stability selection.}
}
\description{
Computer the consistency lower bound
}
