% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{l0}
\alias{l0}
\title{Feature selection subroutine}
\usage{
l0(x, y, s, params, ...)
}
\arguments{
\item{x}{Design matrix.}

\item{y}{Outcome vector.}

\item{s}{Regression method. Current options are \code{"lasso"} or
\code{"boost"}.}

\item{params}{Optional list of parameters to use when \code{s = "boost"}.}

\item{...}{Extra parameters to be passed to the feature selection subroutine.}
}
\description{
This function fits a potentially sparse supervised learning model and returns
a bit vector indicating which features were selected.
}
