% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shah_ss.R
\name{minD}
\alias{minD}
\title{CPSS upper bound}
\usage{
minD(theta, B, r = c(-1/2, -1/4))
}
\arguments{
\item{theta}{Low rate threshold.}

\item{B}{Number of complementary pairs for stability selection.}

\item{r}{Of r-concavity fame.}
}
\description{
Compute the min-D factor of Shah & Samworth's Eq. 8 (2013). Code taken
verbatim from Rajen Shah's personal website:
\url{http://www.statslab.cam.ac.uk/~rds37/papers/r_concave_tail.R}.
}
