% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shah_ss.R
\name{r.TailProbs}
\alias{r.TailProbs}
\title{CPSS utility functions}
\usage{
r.TailProbs(eta, B, r)
}
\description{
These functions are taken verbatim from Rajen Shah's personal website:
\url{http://www.statslab.cam.ac.uk/~rds37/papers/r_concave_tail.R}.
}
