% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sub_loop}
\alias{sub_loop}
\title{Complementary pairs subsampling loop}
\usage{
sub_loop(b, i, j, x, z_t, s, params, ...)
}
\arguments{
\item{b}{Subsample index.}

\item{i}{First foreground variable index.}

\item{j}{Second foreground variable index.}

\item{x}{Matrix of foreground variables.}

\item{z_t}{Intersection of iteration-t known non-descendants for foreground
variables \code{i} and \code{j}.}

\item{s}{Regression method. Current options are \code{"lasso"} or
\code{"boost"}.}

\item{params}{Optional list of parameters to use when \code{s = "boost"}.}

\item{...}{Extra parameters to be passed to the feature selection subroutine.}
}
\description{
This function executes one loop of the model quartet for a given pair of
foreground variables and records any disconnections and/or (de)activations.
}
