% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ss_fn}
\alias{ss_fn}
\title{Infer causal direction using stability selection}
\usage{
ss_fn(df, epsilon, order, rule, B)
}
\arguments{
\item{df}{Table of (de)activation rates.}

\item{epsilon}{Consistency lower bound, as computed by \code{epsilon_fn}.}

\item{order}{Causal order of interest, either \code{"ij"} or \code{"ji"}.}

\item{rule}{Inference rule, either \code{"R1"} or \code{"R2"}.}

\item{B}{Number of complementary pairs to draw for stability selection.}
}
\description{
Infer causal direction using stability selection
}
